Creating a new HWRF branch.

This is to help facilitate code management of the HWRF project with
the DTC, EMC and all other developers.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/HWRF@4946 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ var/gen_be/gen_be_stage2_1dvar.f90
@@ -552,8 +552,8 @@ subroutine da_get_be_inverse( testing_eofs, nk, be, be_inv )
    real                   :: total_var_inv           ! 1 / Total variance of matrix.
    real                   :: cumul_variance          ! Cumulative variance of matrix.
 
-   real                   :: evec(1:nk,1:nk)         ! Eigenvectors.
-   real                   :: eval(1:nk)              ! Eigenvalues.
+   real*8                 :: evec(1:nk,1:nk)         ! Eigenvectors.
+   real*8                 :: eval(1:nk)              ! Eigenvalues.
    real                   :: eval_inv(1:nk)          ! 1 / Eigenvalues.
    real                   :: LamInvET(1:nk,1:nk)     ! ET/sqrt(Eigenvalue).
